Updates landing page and header call to action
Refactors the "Our Solutions" section of the landing page to a more visually appealing and informative layout.

Replaces "Request a Demo" button with "Schedule a Discovery Call" to better reflect the desired user action.

Changes the font to Montserrat across the site to improve the user experience and brand consistency.